Cleaned up Dockerfile and tweaked spigot_init.sh
Closes #5.

The Dockerfile had variables all over the place, they are now cleaned up.

The spigot_init.sh script now checks for an existing jar for that version before rebuilding.  It always rebuilds when the revision is set to latest.  If the revision variable is not set, it is set to latest.  This preserves existing behavior, which is good!

More comments added, and some minor cleanup elsewhere.